refactor: Split loader into separate modules
